refactor: eventhandler 전반 수정
