+++ src/Handler/winningLottoHandler.js
@@ -1,7 +1,9 @@
 import { OPTION } from "../constants/option.js";
 import LottoMachine from "../domain/LottoMachine.js";
+import LottoResult from "../domain/LottoResult.js";
 import WinningLotto from "../domain/WinningLotto.js";
 import bonusNumberValidator from "../validator/BonusNumberValidator.js";
+import WebView from "../view/webView.js";
 
 const winningNumbers = document.querySelectorAll(".input_winningNumber");
 const bonusNumber = document.getElementById("input_bonusNumber");
@@ -34,19 +36,28 @@ const winningLottoHandler = {
   },
 
   validateWinningLotto() {
-    const winningLotto = this.makeWinningLotto();
-    bonusNumberValidator(winningLotto.getNumbers(), Number(bonusNumber.value));
+    const makedLotto = this.makeWinningLotto();
+    bonusNumberValidator(makedLotto.getNumbers(), Number(bonusNumber.value));
     invalidWinningLotto.innerText = "";
+    const winningLotto = WinningLotto(makedLotto, Number(bonusNumber.value));
 
     return winningLotto;
   },
 
-  onClickGameResult(event, resolve) {
+  showModal(lottoList, winningLotto) {
+    const result = new LottoResult(lottoList, winningLotto);
+    const { rank, profit } = result.getResult();
+
+    dialog.showModal();
+    WebView.showGameResult(rank);
+    WebView.showProfit(profit);
+  },
+
+  onClickGameResult(event, lottoList) {
     event.preventDefault();
     try {
       const winningLotto = this.validateWinningLotto();
-      resolve(WinningLotto(winningLotto, Number(bonusNumber.value)));
-      dialog.showModal();
+      this.showModal(lottoList, winningLotto);
     } catch (error) {
       invalidWinningLotto.innerText = error.message;
     }
