+++ src/controller/webController.js
@@ -1,8 +1,7 @@
 import LottoMachine from "../domain/LottoMachine.js";
-import LottoResult from "../domain/LottoResult.js";
-import purchaseClickHandler from "../Handler/purchaseClickHandler.js";
 import retryHandler from "../Handler/retryHandler.js";
 import winningLottoHandler from "../Handler/winningLottoHandler.js";
+import purchaseAmountValidator from "../validator/PurchaseAmountValidator.js";
 import WebView from "../view/webView.js";
 
 const purchaseAmountInput = document.getElementById("input_purchaseAmount");
@@ -11,46 +10,51 @@ const purchaseButton = document.getElementById("purchase_button");
 const winningNumbers = document.querySelectorAll(".input_winningNumber");
 const resultButton = document.getElementById("result_button");
 
-const dialog = document.getElementById("result_dialog");
-const closeDialogButton = document.getElementById("close_dialog");
 const retryButton = document.getElementById("retry_button");
+const invalidPurchaseAmount = document.getElementById("invalid_purchaseAmount");
 
 class WebController {
-  async start() {
+  #lottoList = [];
+
+  start() {
     purchaseAmountInput.focus();
 
-    await this.play();
+    this.play();
     retryButton.addEventListener("click", async () => {
       retryHandler();
     });
   }
 
-  async play() {
-    const lottoList = await this.#setLotto();
-    const winningLotto = await this.#setWinningLotto();
-    this.#getGameResult(lottoList, winningLotto);
+  play() {
+    this.#getPurchaseAmount();
+    this.#getWinningLotto();
   }
 
-  async #setLotto() {
-    const purchaseAmount = await this.#getPurchaseAmount();
-    const lottoMachine = new LottoMachine(purchaseAmount);
-    const lottoList = lottoMachine.makeLottos();
-
-    this.#displayLottoList(lottoList);
-
-    return lottoList;
-  }
-
-  async #getPurchaseAmount() {
-    return new Promise((resolve) => {
-      purchaseButton.addEventListener("click", (event) => {
-        purchaseClickHandler(event, resolve);
-      });
+  #getPurchaseAmount() {
+    purchaseButton.addEventListener("click", (event) => {
+      this.purchaseClickHandler(event);
     });
   }
 
-  #displayLottoList(lottoList) {
-    const lottoNumberList = lottoList.reduce((acc, cur) => {
+  purchaseClickHandler(event) {
+    event.preventDefault();
+    try {
+      purchaseAmountValidator(purchaseAmountInput.value);
+      this.#setLotto();
+      purchaseButton.removeEventListener("click", this.purchaseClickHandler);
+    } catch (error) {
+      invalidPurchaseAmount.innerText = error.message;
+    }
+  }
+
+  #setLotto() {
+    const lottoMachine = new LottoMachine(purchaseAmountInput.value);
+    this.#lottoList = lottoMachine.makeLottos();
+    this.#displayLottoList();
+  }
+
+  #displayLottoList() {
+    const lottoNumberList = this.#lottoList.reduce((acc, cur) => {
       const numbers = cur.getNumbers();
       return [...acc, numbers];
     }, []);
@@ -60,33 +64,13 @@ class WebController {
     WebView.showLottoList(lottoNumberList);
   }
 
-  async #setWinningLotto() {
+  #getWinningLotto() {
     winningNumbers.forEach((input) => {
       input.addEventListener("keydown", winningLottoHandler.onInputKeyDown);
     });
-
-    const winningLotto = await this.#getWinningLotto();
-
-    return winningLotto;
-  }
-
-  #getWinningLotto() {
-    return new Promise((resolve) => {
-      resultButton.addEventListener("click", (event) => {
-        winningLottoHandler.onClickGameResult(event, resolve);
-      });
-      closeDialogButton.addEventListener("click", () => {
-        dialog.close();
-      });
+    resultButton.addEventListener("click", (event) => {
+      winningLottoHandler.onClickGameResult(event, this.#lottoList);
     });
-  }
-
-  #getGameResult(lottoList, winningLotto) {
-    const result = new LottoResult(lottoList, winningLotto);
-    const { rank, profit } = result.getResult();
-
-    WebView.showGameResult(rank);
-    WebView.showProfit(profit);
   }
 }
